fix(LinearPieceWiseEditor): Update widget to use more props and keep less state
The logic embedded in this widget was getting in the way of operations like switching which data array
is active and loading the new points for the other dataset.  Make this widget a simple dislpay of
the points passed into it where all changes should be handled via the onChange callback and the
props should be changed.

This adds a width property to the widget to complement the height
property.  The width and height if set to -1 (the default width) will
cause the SizeHelper to be used to determine the canvas size.  Otherwise
the given size will be used.

This also updates the ColorMapEditor to work with the new
PieceWiseFunctionEditor.

Also updates the API docs for ColorMapEditorWidget and
PieceWiseFunctionEditorWidget.