feat(Containers): Add draggable overlay container
Draggable overlay container can have one or more children of
several types.  It can be re-positioned by dragging the title bar,
or resized by dragging any container edge or corner.  It is
constrained so that no part of it can leave the parent container
where it is placed.

The first use-case for this container is the time probe chart
feature of the TimeFloatImage viewer type in ArcticViewer, and
this topic also improves the way that viewer works.