fix(HistogramSelector): Scoring - uncertainty regions shouldn't overlap
While dragging and when committing to values in the divider popup, make the
uncertainty of the edited divider not overlap with neighbors. When the divider's
value is changed or it's dragged, if it ends up inside another divider's
uncertain region, shrink the other divider's uncertainty.

Also represent uncertainty internally as [0,1] instead of [0,100%]
to make communication with the server easier.