fix(HistogramSelector): Use the HistogramBinHoverProvider
Add handling for highlighting histogram bins that are hovered
with the mouse, and accept hover events from other components.
Don't generate hover events when editing the partition score,
however - it's too distracting. Clear the hover when editing starts.

Add the HistogramBinHoverProvider to the ParallelCoords example,
and update FieldSelector histogram bin styles to match the
HistogramSelector.