fix(MutualInformation): Select status bar text, fix chord selection
Allow status bar text to be selected, and automatically select it
whenever the user clicks a bin or chord that sets the status bar
text. Then the user can Ctrl-C to copy the text.

Use a count for the unique ID - fixes a TODO.

When clicking to create a PMI chord, always run click() before
dblClick() - makes sure the displayed chords are consistent with
the selection/annotation sent to the provider.