WIP: A component to display variables in a graph.
This takes a set of variables (via a FieldProvider), retrieves
mutual information (via a MutualInformationProvider), and generates
a graph of variables, where the distance between nodes is
proportional to the normalized "variation of information."
Variation of information is a distance metric based on mutual
information and joint entropy, since mutual information
is not a metric:

    VofI(a,b) = H(a,b) - MI(a,b)

The normalized distance (between 0 and 1) is:

    d(a,b) = 1 - MI(a,b) / H(a,b) = VofI(a,b) / H(a,b)

which is similar to Jaccard distance.

A slider at the top of the component (which cannot currently be
manipulated except via the keyboard) controls which portions of the
distance matrix are overwritten with large values, allowing the graph
to relax into a shape more indicative of the strongest relationships.