fix(ParallelCoordinates): Add selection resize, single delete
You can resize a selection bar if you drag within 6 pixels of the
ends, or 1/3 of the bar height, if it's too small. This simple
resize works pretty well, but we may want to allow resize when
starting just outside the bar boundaries, in the future.

If you drag a bar totally off the end of the axis, it is deleted.