fix(ConvertProxyProperty): Fix dependency on props with enum domain
Some properties are "dependendents" and should only be shown if some other property takes on a

specific value.  When a property depended on another property with an enumeration domain, there was

an inconsistency where the server provided the dependency using the numeric value of the property,

but gave the data value of that property using the text/label value instead.  By changing the server

to send the numeric value as the property data value, and making this change to compare the

stringified property value, the dependency issue is fixed.