Add onInteraction to VtkRenderer props.
It is an optional function that takes one boolean arg.  The arg is true when a
mouse interaction starts (i.e., mouse down event), false when the interaction
ends (i.e., mouse up event). It just forwards VtkMouseListener.onInteraction
event.

It will be used to extend the paraview websocket lease.